+++ src/Execute-TinyTex_function.ps1
@@ -160,10 +160,9 @@ function Execute-TinyTeX {
     } elseif ( $InstallJaPackages ){
         # equivalent to:
         [String[]] $cmd += @("-e", "library(tinytex);")
-        #[String[]] $cmd += @("-e", "tinytex::tlmgr_install(c('bxjscls', 'luatexja', 'zxjatype', 'zxjafont', 'haranoaji', 'jlreq'));")
-        [String[]] $cmd += @("-e", "tinytex::tlmgr_install(c('bxjscls', 'luatexja', 'haranoaji'));")
-        [String[]] $cmd += @("-e", "tinytex::tlmgr_install(c('jlreq'));")
-        [String[]] $cmd += @("-e", "tinytex::tlmgr_install(c('pgf', 'preview', 'xcolor'));")
+        #[String[]] $cmd += @("-e", """tinytex::tlmgr_install(c('bxjscls', 'luatexja', 'zxjatype', 'zxjafont', 'haranoaji', 'jlreq'));""")
+        [String[]] $cmd += @("-e", """tinytex::tlmgr_install(c('bxjscls', 'luatexja', 'haranoaji', 'jlreq'));""")
+        [String[]] $cmd += @("-e", """tinytex::tlmgr_install(c('pgf', 'preview', 'xcolor'));""")
         [String[]] $cmd += @("-e", "tinytex::tlmgr_path('add');")
         [String[]] $cmd += @("-e", "tinytex::tlmgr_update();")
     } elseif ( $AddPackage ){