+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 - Added [Execute-TinyTeX] `-RemovePackage` option
+- Added [math2tex] physics, siunitx package
 
 ## [7.1.0] - 2024-02-18