+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Added [Execute-TinyTeX] `-RemovePackage` option
+- Added [math2tex] physics, siunitx package
 
 ## [7.1.0] - 2024-02-18